feat(helm): standardize debug output for releases
